Add ack/nack, better errors, & example handlers
Ported ack and nack to the monadic context

Created a ClientErr that was passed back whenever an error file was
needed s.t. a user can now declare an error handler and call it in
combination with publish/ack/nack

Created a error handler and delivery handler examples in main as an
exercise to show how a user might better deal with the CHMQClient API
functions - might want to include some mechanism for a user to pass a
simpler function - might be fine as is.