Add initial modifications for part 2 of spec
Began integrating the fact that queues can now be declared into the CHMQ
server.

I've partially ported publishHandler and have fully ported getHandler.
The other handler still need to be updated (on the serverside)

The clientside has been untouched by changes outside of a couple
necessary changes to get it to temporarily compile. Will be updating
that once the server is finished.